Add avalonia android / desktop / browser / iOS

+++ Templates/Default/Resources.cs
@@ -1,7 +1,13 @@
 foreach (var path in Directory.GetFiles(Project.ResourcesPath, "*.*", SearchOption.AllDirectories))
 {
 	var newPath = path.Replace(Project.ResourcesPath, Project.OutputPath).Replace("Project.Namespace", Project.Namespace);
-	var dirPath = Path.GetDirectoryName(newPath);
-	Directory.CreateDirectory(dirPath);
-	File.Copy(path, newPath, true);
+	Directory.CreateDirectory(Path.GetDirectoryName(newPath));
+
+	// Validate modify time of the file, if the file is not modified, skip copying the file to save time
+	var lastWriteTime = File.GetLastWriteTime(path);
+	if (!File.Exists(newPath) || lastWriteTime != File.GetLastWriteTime(newPath))
+	{
+		File.Copy(path, newPath, true);
+		File.SetLastWriteTime(newPath, lastWriteTime);
+	}
 }
